HAAR-1891: Update following PR comments
