+++ feature.env
@@ -1,6 +1,6 @@
 PORT=3007
 HMPPS_AUTH_URL=http://localhost:9091/auth
-MANAGE_USERS_API_URL=http://localhost:9091
+MANAGE_USERS_API_URL=http://localhost:9091/manage-users-api
 TOKEN_VERIFICATION_API_URL=http://localhost:9091/verification
 TOKEN_VERIFICATION_ENABLED=true
 NODE_ENV=development
