+++ server/data/manageUsersApiClient.test.ts
@@ -2,11 +2,8 @@ import nock from 'nock'
 
 import config from '../config'
 import ManageUsersApiClient from './manageUsersApiClient'
-import TokenStore from './tokenStore'
 
 jest.mock('./tokenStore')
-
-const tokenStore = new TokenStore(null) as jest.Mocked<TokenStore>
 
 const token = { access_token: 'token-1', expires_in: 300 }
 
@@ -16,7 +13,7 @@ describe('manageUsersApiClient', () => {
 
   beforeEach(() => {
     fakeManageUsersApiClient = nock(config.apis.manageUsersApi.url)
-    manageUsersApiClient = new ManageUsersApiClient(tokenStore)
+    manageUsersApiClient = new ManageUsersApiClient()
   })
 
   afterEach(() => {
